Implemented payload that generates header fields

+++ tcp_factory.h
@@ -35,7 +35,7 @@ public:
     struct ev_async factory_async;
     std::list<TcpWorker*> workers;
 
-    uint32_t cumulative_count;
+    uint64_t cumulative_count;
     uint64_t bytes_in;
     uint64_t bytes_out;
